Add support for TI Test Automation Interface
This is an initial integration of TIs Test Automation Interface as a
Boardswarm actuator. At the moment this allows the power to the board to
be turned on and off and optionally set the boot mode of the device.

The TI Test Automation Interface is known to be available (if it is
enabled) through the JTAG interface on the TI SK-AM62 and SK-AM62P-LP
development boards.

A tool called "xdsdfu", which is shipped as part of the TI's Code
Composer Studio IDE, needs to be used to put the device into the correct
mode and configuration:

- Put device into DFU mode:
    xdsdfu -m

- Set configuration to “5” and reset:
    xdsdfu -c 5 -r

Two serial devices will be enumerated, with the 2nd being the automation
interface.

The type of device needs to be specified with a parameter in a provider
node in the server configuration file, for example:

  - name: ti-am62p-xds
    provider: ti-xds-tai
    parameters:
      match:
        udev.ID_PATH: "pci-0000:05:00.3-usb-0:4.2.4:1.0"
      model: "am62pxx_sk-fs"

This interface currently provides the ability to power the device on and
off, which can be configured for a device in the server configuration
file like this:

    modes:
      - name: on
        depends: off
        sequence:
          - match: &pdu
              udev.ID_PATH: "pci-0000:05:00.3-usb-0:4.2.4:1.0"
            parameters:
              mode: mmc
              state: on
      - name: off
        sequence:
          - match: *pdu
            parameters:
              state: off
            stabilisation: 2s

Signed-off-by: Martyn Welch <martyn.welch@collabora.com>